move prefs to separate controller
In service of #348
