--- whatdid.xcodeproj/project.xcworkspace/xcshareddata/IDETemplateMacros.plist
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>FILEHEADER</key>
-  <string> ___PACKAGENAME___?</string>
-</dict>
-</plist>
